refactor(core): extract workspace permission gate into shared module
